Update MyoObjectiveC.mm
added missing gestures also added method to get gesture name by passing pose type. by the way i would appreciate it if you could add a way of getting the firmware version i have tried modifying this to make it work, it compiles and runs but the onPair and onConnect methods don't seem to fire so there is no way to test it. I am also not proficient enough with C++ to be able to figure out why.